# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Qberry API documentation",
        "description": "The company is engaged in the storage of perishable products for customers around the world. With refrigeration capacity in 6 locations, flexible booking system and excellent service, the company has won the trust of large corporations.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/booking-blocks-by-location": {
            "post": {
                "tags": [
                    "Get blocks by location"
                ],
                "summary": "Get all available blocks by using data filtering",
                "operationId": "aa91d7d2d88ac36751743484ea117bd0",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "description": "ID location",
                        "required": true,
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    },
                    {
                        "name": "date_start",
                        "in": "path",
                        "description": "Block lease start date. The start date cannot be less than the current date.",
                        "required": true,
                        "examples": {
                            "int": {
                                "summary": "date starts",
                                "value": "2022-11-04"
                            }
                        }
                    },
                    {
                        "name": "date_end",
                        "in": "path",
                        "description": "Block lease ends date",
                        "required": true,
                        "examples": {
                            "int": {
                                "summary": "date ends",
                                "value": "2022-11-20"
                            }
                        }
                    },
                    {
                        "name": "volume",
                        "in": "path",
                        "description": "Volume of products in m3",
                        "required": true,
                        "examples": {
                            "int": {
                                "summary": "volume",
                                "value": "2"
                            }
                        }
                    },
                    {
                        "name": "temperature",
                        "in": "path",
                        "description": "Refrigerator temperature. The temperature cannot be more than 0 degrees Celsius. If you need to specify temperature 0 - specify -2",
                        "required": true,
                        "examples": {
                            "int": {
                                "summary": "temperature",
                                "value": "-3"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "location": {
                                                "description": "ID location",
                                                "type": "integer"
                                            },
                                            "date_start": {
                                                "description": "Block lease start date. The start date cannot be less than the current date.",
                                                "type": "string"
                                            },
                                            "date_end": {
                                                "description": "Block lease ends date",
                                                "type": "string"
                                            },
                                            "volume": {
                                                "description": "Volume of products in m3",
                                                "type": "integer"
                                            },
                                            "temperature": {
                                                "description": "Refrigerator temperature. The temperature cannot be more than 0 degrees Celsius. If you need to specify temperature 0 - specify -2",
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "location": 1,
                                    "date_start": "2022-11-01 12:00:00",
                                    "date_end": "2022-11-20 12:00:00",
                                    "volume": 2,
                                    "temperature": -3
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "you_need_blocks": {
                                            "type": "number",
                                            "example": "2"
                                        },
                                        "we_have_block": {
                                            "type": "number",
                                            "example": "2"
                                        },
                                        "start_data": {
                                            "type": "string",
                                            "example": "2022-11-01 12:00:00"
                                        },
                                        "end_data": {
                                            "type": "string",
                                            "example": "2022-11-02 12:00:00"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "Wilmington"
                                        },
                                        "sum": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "10"
                                        },
                                        "hash": {
                                            "type": "string",
                                            "example": "$2y$10$ch/XPMjQ0ZAEAibE.ko5iOSOUFcLi4h2K.bDfwac8HpFhd/QsN3F"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "id"
                                                    },
                                                    "fridge_id": {
                                                        "type": "number",
                                                        "example": "23"
                                                    },
                                                    "length": {
                                                        "type": "number",
                                                        "example": "2"
                                                    },
                                                    "width": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "height": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "volume": {
                                                        "type": "number",
                                                        "example": "2"
                                                    },
                                                    "status": {
                                                        "type": "number",
                                                        "example": "0"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Описание ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}